feat: add types to content and dialog
