+++ src/components/signinList.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { useIntl } from "react-intl";
 import { TAB_STATE } from "@pages/popup/constants";
-import { SigninCard, ISignin } from "@components/signinCard";
+import { SigninCard } from "@components/signinCard";
 import { Loader } from "@components/ui";
-import { IMessage } from "@config/types";
+import { IMessage, ISignin } from "@config/types";
 
 interface IResourceSignin {
   index: number;
