+++ src/pages/dialog/index.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { createRoot } from "react-dom/client";
 import "@pages/popup/index.css";
 import "@assets/styles/tailwind.css";
